+++ src/GitVersionCore/Extensions/GitVersionInformationResources/GitVersionInformationGenerator.cs
@@ -5,11 +5,6 @@ using GitVersion.OutputVariables;
 
 namespace GitVersion.Extensions.GitVersionInformationResources
 {
-    public interface IGitVersionInformationGenerator
-    {
-        void Generate(VersionVariables variables, FileWriteInfo writeInfo);
-    }
-
     public class GitVersionInformationGenerator : IGitVersionInformationGenerator
     {
         private readonly IFileSystem fileSystem;
