+++ src/GitVersionCore/Extensions/GitVersionInformationResources/IGitVersionInformationGenerator.cs
@@ -1,0 +1,9 @@
+using GitVersion.OutputVariables;
+
+namespace GitVersion.Extensions.GitVersionInformationResources
+{
+    public interface IGitVersionInformationGenerator
+    {
+        void Generate(VersionVariables variables, FileWriteInfo writeInfo);
+    }
+}
