+++ src/GitVersionCore/Extensions/IWixVersionFileUpdater.cs
@@ -1,0 +1,10 @@
+using System;
+using GitVersion.OutputVariables;
+
+namespace GitVersion.Extensions
+{
+    public interface IWixVersionFileUpdater : IDisposable
+    {
+        string Update(VersionVariables variables, string workingDirectory);
+    }
+}
