+++ src/GitVersionCore/Extensions/WixVersionFileUpdater.cs
@@ -7,11 +7,6 @@ using GitVersion.OutputVariables;
 
 namespace GitVersion.Extensions
 {
-    public interface IWixVersionFileUpdater : IDisposable
-    {
-        string Update(VersionVariables variables, string workingDirectory);
-    }
-
     public class WixVersionFileUpdater : IWixVersionFileUpdater
     {
         private readonly IFileSystem fileSystem;
