+++ src/GitVersionCore/Model/FileWriteInfo.cs
@@ -1,4 +1,4 @@
-﻿namespace GitVersion
+namespace GitVersion
 {
     public sealed class FileWriteInfo
     {
