+++ src/GitVersionExe.Tests/GitVersionExe.Tests.csproj
@@ -10,6 +10,7 @@
         <PackageReference Include="Mono.Cecil" Version="0.11.2"></PackageReference>
     </ItemGroup>
     <ItemGroup>
+        <ProjectReference Include="..\GitVersionCore.Tests\GitVersionCore.Tests.csproj" />
         <ProjectReference Include="..\GitVersionCore\GitVersionCore.csproj" />
         <ProjectReference Include="..\GitVersionExe\GitVersionExe.csproj" />
     </ItemGroup>
